test(platform-browser): consolidate BrowserCrypto coverage
